update self updater files

+++ nuinstall_lib.php
@@ -12,6 +12,27 @@ class nuinstall {
 	var $warnings        = array();
 	var $removeColumns   = false;
 	var $removeIndexes   = false;
+	var $overrideSetup   = false;
+	var $initResult	     = 'UNKNOWN';
+
+	function checkInstall() {
+
+		if ( $this->checkDatabaseConnection() ) {
+			if ( $this->checkDatabaseExists() ) {
+				if ( $this->checkTableExists('zzzsys_setup') ) {
+					$this->initResult = 'OK';
+					$this->overrideSetup = true;	
+				} else {
+					$this->initResult = 'SCHEMA_INCOMPLETE';
+					$this->overrideSetup = true;
+				}
+			} else {
+				$this->initResult = 'DATABASE_NOT_CREATED';
+			}	
+		} else {
+			$this->initResult = 'CANNOT_CONNECT_TO_SERVER';
+		}
+	}
 
         function setDB($DBHost, $DBName, $DBUserID, $DBPassWord) {
                 $this->DB['DBHost']       = $DBHost;
@@ -19,6 +40,56 @@ class nuinstall {
                 $this->DB['DBUserID']     = $DBUserID;
                 $this->DB['DBPassWord']   = $DBPassWord;
         }
+
+	function checkDatabaseConnection() {
+
+                $DBHost         = $this->DB['DBHost'];
+                $DBUserID       = $this->DB['DBUserID'];
+                $DBPassWord     = $this->DB['DBPassWord'];
+                $DBName         = $this->DB['DBName'];
+
+		$con = false;
+                $con = mysql_connect($DBHost,$DBUserID,$DBPassWord);
+                if ($con === false) {
+			return false;
+                } else {
+			return true;
+		}
+        }
+
+	function checkDatabaseExists() {
+
+                $DBHost         = $this->DB['DBHost'];
+                $DBUserID       = $this->DB['DBUserID'];
+                $DBPassWord     = $this->DB['DBPassWord'];
+                $DBName         = $this->DB['DBName'];
+
+                $con = mysql_connect($DBHost,$DBUserID,$DBPassWord);
+		
+		$sdb = mysql_select_db($DBName,$con);
+
+		if ( !$sdb ) {
+			return false;
+		} else {
+			return true;
+		}	
+        }
+
+	function checkTableExists($table) {
+
+		$DBName                 = $this->DB['DBName'];
+		$dbInfo                 = $this->DB;
+		$dbInfo['DBName']       = "information_schema";
+		$sql                    = "SELECT table_name FROM tables WHERE table_schema = '$DBName' AND table_name = '$table'";
+		$rs                     = $this->runQuery($sql, $dbInfo);
+		$num                    = mysql_num_rows($rs);
+
+		if ($num != 1 ) {
+			return false;
+		} else {
+			return true;	
+		}
+	}
 
 	function importTemplate() {
 
@@ -152,12 +223,16 @@ class nuinstall {
 
 	function deleteNubuilderInfo($template_table, $real_table) {
 
-		// delete existing nuBuilder info in zzsys tables
-                $this->addDisplay("<b>Deleting nuBuilder info from table</b><br>");
-                $id        = $real_table."_id";
-                $clean_sql = "DELETE FROM $real_table WHERE $id IN (SELECT $id FROM $template_table) ";
-                $this->addDisplay($clean_sql."<br>");
-                $this->runQuery($clean_sql, $this->DB);	
+		if ( $this->overrideSetup == false && $real_table == 'zzzsys_setup' ) {
+			$this->addDisplay("<b>Skipping Delete zzzsys_table</b><br>");
+		} else {
+			// delete existing nuBuilder info in zzsys tables
+        	        $this->addDisplay("<b>Deleting nuBuilder info from table</b><br>");
+                	$id        = $real_table."_id";
+	                $clean_sql = "DELETE FROM $real_table WHERE $id IN (SELECT $id FROM $template_table) ";
+                	$this->addDisplay($clean_sql."<br>");
+                	$this->runQuery($clean_sql, $this->DB);	
+		}
 	}
 
 	function loopColumns($template_table, $real_table) {
@@ -309,27 +384,32 @@ class nuinstall {
 		}
                 $colum_order2 = implode(", ", $colum_order);
 
-		// do insert
-                $this->addDisplay("<b>Inserting nuBuilder info into table</b><br>");
-                $insert = "INSERT INTO `$real_table` SELECT $colum_order2 FROM `$template_table` ";
-                $this->runQuery($insert, $this->DB);
-		$this->addDisplay($insert."<br>");
+		if ( $this->overrideSetup == false && $real_table == 'zzzsys_setup' ) {
+			$this->addDisplay("<b>Skipping Insert into zzzsys_table</b><br>");
+		} else {
 
-		//debug errors
-		if ( $this->lastSQLerror != "" ) {
-			$this->addDisplay("<h3>".$this->lastSQLerror."</h3>");
+			// do insert
+                	$this->addDisplay("<b>Inserting nuBuilder info into table</b><br>");
+                	$insert = "INSERT INTO `$real_table` SELECT $colum_order2 FROM `$template_table` ";
+                	$this->runQuery($insert, $this->DB);
 			$this->addDisplay($insert."<br>");
-			$this->addDisplay("<pre>");
-	                $this->addDisplay($real_table."<br>");
-        	        $this->addDisplay(count($real_table_columns)."<br>");
-                	$this->addDisplay(print_r($real_table_columns, true));
-	                $this->addDisplay($template_table."<br>");
-        	        $this->addDisplay(count($template_table_columns)."<br>");
-                	$this->addDisplay(print_r($template_table_columns,true));
-	                $this->addDisplay(count($colum_order)."<br>");
-        	        $this->addDisplay(print_r($colum_order,true));
-                	$this->addDisplay("</pre>");
-			$this->addDisplay("<hr>");
+
+			//debug errors
+			if ( $this->lastSQLerror != "" ) {
+				$this->addDisplay("<h3>".$this->lastSQLerror."</h3>");
+				$this->addDisplay($insert."<br>");
+				$this->addDisplay("<pre>");
+	                	$this->addDisplay($real_table."<br>");
+        	        	$this->addDisplay(count($real_table_columns)."<br>");
+                		$this->addDisplay(print_r($real_table_columns, true));
+	                	$this->addDisplay($template_table."<br>");
+        	        	$this->addDisplay(count($template_table_columns)."<br>");
+                		$this->addDisplay(print_r($template_table_columns,true));
+	                	$this->addDisplay(count($colum_order)."<br>");
+        	        	$this->addDisplay(print_r($colum_order,true));
+                		$this->addDisplay("</pre>");
+				$this->addDisplay("<hr>");
+			}
 		}
 		
 	}
